pyproject.toml
[tool.poetry]
name = "janelle"
version = "0.1.0"
description = "Discord moderation & management bot for IdleRPG"
authors = ["Jens Reidel <adrian@travitia.xyz>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
"discord.py-stubs" = "^1.5.0"
discord-ext-typed-commands = "^1.0.2"

[tool.poetry.dev-dependencies]
black = { version = "^20.8b1", allow-prereleases = true }

[tool.poetry.scripts]
bot = "bot:main"

[tool.flake8]
max-line-length = 88

[tool.isort]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
force_grid_wrap = 0
line_length = 88
lines_between_types = 1

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
